requirements.txt
# Backend dependencies
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic>=2.3.0
pydantic-settings>=2.0.0
supabase>=2.0.0
httpx>=0.24.1
python-dotenv>=1.0.0
PyJWT>=2.8.0
google-generativeai>=0.1.0
